Add workarounds for Linux ACLs
